ui/streamlit_app/requirements.txt
streamlit==1.38.0
plotly==5.24.1
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
supabase==2.0.3
